directly read from KML skipping text file; rename parser to avoid conflict
